Switch to using ez_setup.py, and simplify setup.py, following recent changes in Astropy core.
